+++ inference-engine/tests/functional/plugin/CMakeLists.txt
@@ -19,3 +19,4 @@ endif()
 if (ENABLE_MYRIAD)
     add_subdirectory(myriad)
 endif()
+
