+++ inference-engine/tests/functional/plugin/shared/include/single_layer_tests/one_hot.hpp
@@ -1,30 +1,15 @@
-// Copyright (C) 2020 Intel Corporation
+// Copyright (C) 2019 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
 
 #pragma once
 
-#include "shared_test_classes/base/layer_test_utils.hpp"
-#include "ngraph_functions/builders.hpp"
-
-using oneHotParams = typename std::tuple<
-    int64_t,                    // axis
-    size_t,                     // depth
-    float,                      // on_value
-    float,                      // off_value
-    std::vector<size_t>,        // Input shapes
-    InferenceEngine::Precision, // Net precision
-    LayerTestsUtils::TargetDevice // Target device name
-    >;
+#include "shared_test_classes/single_layer/one_hot.hpp"
 
 namespace LayerTestsDefinitions {
 
-class OnehotLayerTest : public testing::WithParamInterface<oneHotParams>, public LayerTestsUtils::LayerTestsCommon {
-  public:
-    static std::string getTestCaseName(testing::TestParamInfo<oneHotParams> obj);
-
-  protected:
-    void SetUp() override;
+TEST_P(OnehotLayerTest, CompareWithRefs) {
+    Run();
 };
 
 } // namespace LayerTestsDefinitions
