+++ inference-engine/tests/functional/plugin/shared/include/single_layer_tests/reorg_yolo.hpp
@@ -13,4 +13,3 @@ TEST_P(ReorgYoloLayerTest, CompareWithRefs) {
 };
 
 } // namespace LayerTestsDefinitions
-
