+++ inference-engine/tests/functional/shared_test_classes/src/single_layer/logical.cpp
