+++ inference-engine/tests/functional/shared_test_classes/src/single_layer/reorg_yolo.cpp
@@ -32,4 +32,3 @@ void ReorgYoloLayerTest::SetUp() {
 }
 
 } // namespace LayerTestsDefinitions
-
