+++ inference-engine/tests/functional/shared_test_classes/src/single_layer/squared_difference.cpp
@@ -2,25 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include <tuple>
-#include <string>
-#include <vector>
-#include <memory>
-#include <functional>
-#include <functional_test_utils/skip_tests_config.hpp>
-
-#include "ie_core.hpp"
-#include "ngraph_functions/utils/ngraph_helpers.hpp"
-
-#include "ngraph/op/squared_difference.hpp"
-
-#include "common_test_utils/common_utils.hpp"
-#include "functional_test_utils/precision_utils.hpp"
-#include "functional_test_utils/blob_utils.hpp"
-#include "functional_test_utils/plugin_cache.hpp"
-#include "shared_test_classes/base/layer_test_utils.hpp"
-
-#include "single_layer_tests/squared_difference.hpp"
+#include "shared_test_classes/single_layer/squared_difference.hpp"
 
 namespace LayerTestsDefinitions {
 
@@ -43,8 +25,7 @@ void SquaredDifferenceLayerTest::SetUp() {
     auto ngPrc = FuncTestUtils::PrecisionUtils::convertIE2nGraphPrc(netPrecision);
     auto params = ngraph::builder::makeParams(ngPrc, {inputShapes});
 
-    auto paramOuts = ngraph::helpers::convert2OutputVector(
-            ngraph::helpers::castOps2Nodes<ngraph::op::Parameter>(params));
+    auto paramOuts = ngraph::helpers::convert2OutputVector(ngraph::helpers::castOps2Nodes<ngraph::op::Parameter>(params));
     IE_ASSERT(paramOuts.size() == 2);
     const auto squared_difference = std::make_shared<ngraph::opset1::SquaredDifference>(paramOuts.at(0), paramOuts.at(1));
 
@@ -52,9 +33,5 @@ void SquaredDifferenceLayerTest::SetUp() {
     results.push_back(std::make_shared<ngraph::opset1::Result>(squared_difference));
     function = std::make_shared<ngraph::Function>(results, params, "squared_difference");
 }
-
-TEST_P(SquaredDifferenceLayerTest, CompareWithRefs) {
-    Run();
-};
 
 }  // namespace LayerTestsDefinitions
