+++ inference-engine/tests/ngraph_helpers/ngraph_functions/src/variadic_split.cpp
@@ -10,11 +10,11 @@
 namespace ngraph {
 namespace builder {
         std::shared_ptr<ngraph::Node> makeVariadicSplit(const ngraph::Output<Node> &in,
-                                                        const std::vector<int32_t> numSplits,
+                                                        const std::vector<size_t> numSplits,
                                                         size_t axis) {
             auto splitAxisOp = std::make_shared<ngraph::opset3::Constant>(element::u64, ngraph::Shape{},
                                                                           std::vector<size_t>{axis});
-            auto numSplit = std::make_shared<ngraph::opset3::Constant>(element::i64, ngraph::Shape{numSplits.size()},
+            auto numSplit = std::make_shared<ngraph::opset3::Constant>(element::u64, ngraph::Shape{numSplits.size()},
                                                                        numSplits);
             auto VariadicSplitNode = std::make_shared<ngraph::opset3::VariadicSplit>(in, splitAxisOp, numSplit);
             return VariadicSplitNode;
